u-boot-toradex: Update both u-boot and fw-utils to version 2019.07
Use u-boot-toradex-fw-utils recipe from meta-toradex-nxp layer [1]

1 - https://git.toradex.com/cgit/meta-toradex-nxp.git/tree/recipes-bsp/u-boot/u-boot-toradex-fw-utils_2019.07.bb

+++ recipes-bsp/u-boot/u-boot-toradex-fw-utils_2019.07.bb
@@ -1,30 +1,51 @@
-require recipes-bsp/u-boot/u-boot-toradex_${PV}.inc
+require u-boot-toradex-common.inc
 
 SUMMARY = "U-Boot bootloader fw_printenv/setenv utilities"
-
-SRC_URI += "file://0001-tools-fix-cross-compiling-tools-when-HOSTCC-is-overr.patch \
-            file://fw_env.config \
-"
-
 DEPENDS += "mtd-utils"
+
+SRC_URI += " \
+    file://fw_env.config \
+    file://fw_unlock_mmc.sh \
+"
 
 INSANE_SKIP_${PN} = "already-stripped"
 EXTRA_OEMAKE_class-target = 'CROSS_COMPILE=${TARGET_PREFIX} CC="${CC} ${CFLAGS} ${LDFLAGS}" HOSTCC="${BUILD_CC} ${BUILD_CFLAGS} ${BUILD_LDFLAGS}" V=1'
-EXTRA_OEMAKE_class-cross = 'ARCH=${TARGET_ARCH} CC="${CC} ${CFLAGS} ${LDFLAGS}" V=1'
+EXTRA_OEMAKE_class-cross = 'HOSTCC="${CC} ${CFLAGS} ${LDFLAGS}" V=1'
 
 inherit uboot-config
 
 do_compile () {
     oe_runmake ${UBOOT_MACHINE}
-    oe_runmake env
+    oe_runmake envtools
 }
 
 do_install () {
     install -d ${D}${base_sbindir}
     install -d ${D}${sysconfdir}
     install -m 755 ${S}/tools/env/fw_printenv ${D}${base_sbindir}/fw_printenv
-    install -m 755 ${S}/tools/env/fw_printenv ${D}${base_sbindir}/fw_setenv
-    install -m 0644 ${WORKDIR}/fw_env.config ${D}${sysconfdir}/fw_env.config
+    ln -s fw_printenv ${D}${base_sbindir}/fw_setenv
+
+    if [ -e ${WORKDIR}/fw_env.config ]; then
+        install -m 0644 ${WORKDIR}/fw_env.config ${D}${sysconfdir}/fw_env.config
+    else
+        install -m 0644 ${S}/tools/env/fw_env.config ${D}${sysconfdir}/fw_env.config
+    fi
+}
+
+install_unlock_emmc() {
+    install -Dm 0644 ${WORKDIR}/fw_unlock_mmc.sh ${D}${sysconfdir}/profile.d/fw_unlock_mmc.sh
+}
+
+do_install_append_apalis-imx6() {
+    install_unlock_emmc
+}
+
+do_install_append_colibri-imx6() {
+    install_unlock_emmc
+}
+
+do_install_append_colibri-imx7-emmc() {
+    install_unlock_emmc
 }
 
 do_install_class-cross () {
